sports_betting/requirements.txt
requests>=2.31.0
pandas>=2.2.3
numpy>=2.0.0
scipy>=1.13.0
xgboost>=2.1.0
scikit-learn>=1.5.0
flask>=3.0.0
schedule>=1.2.1
python-dotenv>=1.0.1
